+++ src/bot/index.ts
@@ -96,36 +96,10 @@ const MAX_OUTBOUND_GENERATED_IMAGES = 4;
 const RESPONSE_DELIVERY_MODE = resolveResponseDeliveryMode();
 export const getResponseDeliveryModeForTest = (): string => RESPONSE_DELIVERY_MODE;
 
-// =============================================================================
-// Retry store — keeps retry info for the Retry button on errors
-// =============================================================================
-export const RETRY_BTN_PREFIX = 'retry_prompt_';
-const MAX_RETRY_STORE_SIZE = 100;
-
-export interface RetryInfo {
-    /** 'resend' = prompt was never sent, re-inject it; 'click-retry' = prompt already sent, click Antigravity retry button */
-    type: 'resend' | 'click-retry';
-    /** Original prompt text (only used for 'resend') */
-    prompt?: string;
-}
-
-const retryStore = new Map<string, RetryInfo>();
-
-function storeRetry(key: string, info: RetryInfo): void {
-    if (retryStore.size >= MAX_RETRY_STORE_SIZE) {
-        const firstKey = retryStore.keys().next().value;
-        if (firstKey !== undefined) retryStore.delete(firstKey);
-    }
-    retryStore.set(key, info);
-}
-
-export function getRetryInfo(key: string): RetryInfo | undefined {
-    return retryStore.get(key);
-}
-
-export function deleteRetryInfo(key: string): void {
-    retryStore.delete(key);
-}
+// Retry store (extracted to avoid circular dependency with interactionCreateHandler)
+import { RETRY_BTN_PREFIX, storeRetry, RetryInfo } from '../services/retryStore';
+export { RETRY_BTN_PREFIX, RetryInfo } from '../services/retryStore';
+export { getRetryInfo, deleteRetryInfo } from '../services/retryStore';
 
 export function createSerialTaskQueueForTest(queueName: string, traceId: string): (task: () => Promise<void>, label?: string) => Promise<void> {
     let queue: Promise<void> = Promise.resolve();
@@ -207,7 +181,7 @@ async function sendPromptToAntigravity(
     const sendErrorWithRetry = (
         title: string,
         description: string,
-        retryInfo: RetryInfo,
+        retryInfo: Omit<RetryInfo, 'createdAt'>,
     ): Promise<void> => enqueueGeneral(async () => {
         if (!channel) return;
         const retryId = message.id;
@@ -914,89 +888,97 @@ export const startBot = async () => {
             }
         },
         handleRetry: async (interaction, retryInfo) => {
-            // Resolve CDP via workspace binding (same flow as template handler)
-            const channelId = interaction.channelId;
-            const workspacePath = wsHandler.getWorkspaceForChannel(channelId);
+            try {
+                // Resolve CDP via workspace binding (same flow as template handler)
+                const channelId = interaction.channelId;
+                const workspacePath = wsHandler.getWorkspaceForChannel(channelId);
 
-            let cdp: CdpService | null = null;
-            if (workspacePath) {
-                try {
-                    cdp = await bridge.pool.getOrConnect(workspacePath);
-                    const dirName = bridge.pool.extractDirName(workspacePath);
-                    bridge.lastActiveWorkspace = dirName;
-                    bridge.lastActiveChannel = interaction.channel;
-                    registerApprovalWorkspaceChannel(bridge, dirName, interaction.channel as any);
-                    const session = chatSessionRepo.findByChannelId(channelId);
-                    if (session?.displayName) {
-                        registerApprovalSessionChannel(bridge, dirName, session.displayName, interaction.channel as any);
+                let cdp: CdpService | null = null;
+                if (workspacePath) {
+                    try {
+                        cdp = await bridge.pool.getOrConnect(workspacePath);
+                        const dirName = bridge.pool.extractDirName(workspacePath);
+                        bridge.lastActiveWorkspace = dirName;
+                        bridge.lastActiveChannel = interaction.channel;
+                        registerApprovalWorkspaceChannel(bridge, dirName, interaction.channel as any);
+                        const session = chatSessionRepo.findByChannelId(channelId);
+                        if (session?.displayName) {
+                            registerApprovalSessionChannel(bridge, dirName, session.displayName, interaction.channel as any);
+                        }
+                        ensureApprovalDetector(bridge, cdp, dirName, client);
+                    } catch (e: any) {
+                        await interaction.followUp({
+                            content: `Failed to connect to workspace: ${e.message}`,
+                            flags: MessageFlags.Ephemeral,
+                        });
+                        return;
                     }
-                    ensureApprovalDetector(bridge, cdp, dirName, client);
-                } catch (e: any) {
-                    await interaction.followUp({
-                        content: `Failed to connect to workspace: ${e.message}`,
-                        flags: MessageFlags.Ephemeral,
-                    });
-                    return;
-                }
-            } else {
-                cdp = getCurrentCdp(bridge);
-            }
-
-            if (!cdp) {
-                await interaction.followUp({
-                    content: t('Not connected to CDP. Please connect to a project first.'),
-                    flags: MessageFlags.Ephemeral,
-                });
-                return;
-            }
-
-            if (retryInfo.type === 'click-retry') {
-                // Post-injection error: prompt already in Antigravity, click its retry button
-                const clickResult = await cdp.call('Runtime.evaluate', {
-                    expression: RESPONSE_SELECTORS.CLICK_RETRY_BUTTON,
-                    returnByValue: true,
-                    awaitPromise: false,
-                }).catch(() => null);
-                const value = clickResult?.result?.value;
-
-                if (value?.ok) {
-                    await interaction.followUp({
-                        content: t('🔄 Clicked Antigravity retry button. Response will regenerate.'),
-                    }).catch(logger.error);
                 } else {
-                    await interaction.followUp({
-                        content: t('Antigravity retry button not found. Please retry manually from Antigravity.'),
-                        flags: MessageFlags.Ephemeral,
-                    }).catch(logger.error);
+                    cdp = getCurrentCdp(bridge);
                 }
-            } else {
-                // Pre-injection error: prompt was never sent, re-inject it
-                if (!retryInfo.prompt) {
+
+                if (!cdp) {
                     await interaction.followUp({
-                        content: t('Retry data not found. The prompt may have expired.'),
+                        content: t('Not connected to CDP. Please connect to a project first.'),
                         flags: MessageFlags.Ephemeral,
                     });
                     return;
                 }
 
-                const followUp = await interaction.followUp({
-                    content: t('🔄 Retrying...'),
-                });
+                if (retryInfo.type === 'click-retry') {
+                    // Post-injection error: prompt already in Antigravity, click its retry button
+                    const clickResult = await cdp.call('Runtime.evaluate', {
+                        expression: RESPONSE_SELECTORS.CLICK_RETRY_BUTTON,
+                        returnByValue: true,
+                        awaitPromise: false,
+                    }).catch(() => null);
+                    const value = clickResult?.result?.value;
 
-                if (followUp instanceof Message) {
-                    await promptDispatcher.send({
-                        message: followUp,
-                        prompt: retryInfo.prompt,
-                        cdp,
-                        inboundImages: [],
-                        options: {
-                            chatSessionService,
-                            chatSessionRepo,
-                            channelManager,
-                            titleGenerator,
-                        },
+                    if (value?.ok) {
+                        await interaction.followUp({
+                            content: t('🔄 Clicked Antigravity retry button. Response will regenerate.'),
+                        }).catch(logger.error);
+                    } else {
+                        await interaction.followUp({
+                            content: t('Antigravity retry button not found. Please retry manually from Antigravity.'),
+                            flags: MessageFlags.Ephemeral,
+                        }).catch(logger.error);
+                    }
+                } else {
+                    // Pre-injection error: prompt was never sent, re-inject it
+                    if (!retryInfo.prompt) {
+                        await interaction.followUp({
+                            content: t('Retry data not found. The prompt may have expired.'),
+                            flags: MessageFlags.Ephemeral,
+                        });
+                        return;
+                    }
+
+                    const followUp = await interaction.followUp({
+                        content: t('🔄 Retrying...'),
                     });
+
+                    if (followUp instanceof Message) {
+                        await promptDispatcher.send({
+                            message: followUp,
+                            prompt: retryInfo.prompt,
+                            cdp,
+                            inboundImages: [],
+                            options: {
+                                chatSessionService,
+                                chatSessionRepo,
+                                channelManager,
+                                titleGenerator,
+                            },
+                        });
+                    }
                 }
+            } catch (error: any) {
+                logger.error('[Retry] handleRetry failed:', error);
+                await interaction.followUp({
+                    content: t('An error occurred during retry. Please try again.'),
+                    flags: MessageFlags.Ephemeral,
+                }).catch(logger.error);
             }
         },
     }));
