+++ src/events/interactionCreateHandler.ts
@@ -1,6 +1,4 @@
 import {
-    ActionRowBuilder,
-    ButtonBuilder,
     ButtonInteraction,
     ChatInputCommandInteraction,
     EmbedBuilder,
@@ -11,7 +9,8 @@ import {
 
 import { t } from '../utils/i18n';
 import { logger } from '../utils/logger';
-import { RETRY_BTN_PREFIX, getRetryInfo, deleteRetryInfo, RetryInfo } from '../bot/index';
+import { RETRY_BTN_PREFIX, getRetryInfo, deleteRetryInfo, RetryInfo } from '../services/retryStore';
+import { disableAllButtons } from '../ui/buttonUtils';
 import { TEMPLATE_BTN_PREFIX, parseTemplateButtonId } from '../ui/templateUi';
 import {
     AUTOACCEPT_BTN_OFF,
@@ -127,27 +126,7 @@ export function createInteractionCreateHandler(deps: InteractionCreateHandlerDep
                                 .addFields({ name: 'Action History', value: historyText, inline: false })
                                 .setTimestamp();
 
-                            const disabledRows = interaction.message.components
-                                .map((row) => {
-                                    const rowAny = row as any;
-                                    if (!Array.isArray(rowAny.components)) return null;
-
-                                    const nextRow = new ActionRowBuilder<ButtonBuilder>();
-                                    const disabledButtons = rowAny.components
-                                        .map((component: any) => {
-                                            const componentType = component?.type ?? component?.data?.type;
-                                            if (componentType !== 2) return null;
-                                            const payload = typeof component?.toJSON === 'function'
-                                                ? component.toJSON()
-                                                : component;
-                                            return ButtonBuilder.from(payload).setDisabled(true);
-                                        })
-                                        .filter((button: ButtonBuilder | null): button is ButtonBuilder => button !== null);
-                                    if (disabledButtons.length === 0) return null;
-                                    nextRow.addComponents(...disabledButtons);
-                                    return nextRow;
-                                })
-                                .filter((row): row is ActionRowBuilder<ButtonBuilder> => row !== null);
+                            const disabledRows = disableAllButtons(interaction.message.components);
 
                             await interaction.update({
                                 embeds: [updatedEmbed],
@@ -267,27 +246,7 @@ export function createInteractionCreateHandler(deps: InteractionCreateHandlerDep
                     deleteRetryInfo(retryId);
 
                     // Disable the retry button after click
-                    const disabledRows = interaction.message.components
-                        .map((row) => {
-                            const rowAny = row as any;
-                            if (!Array.isArray(rowAny.components)) return null;
-                            const nextRow = new ActionRowBuilder<ButtonBuilder>();
-                            const disabledButtons = rowAny.components
-                                .map((component: any) => {
-                                    const componentType = component?.type ?? component?.data?.type;
-                                    if (componentType !== 2) return null;
-                                    const payload = typeof component?.toJSON === 'function'
-                                        ? component.toJSON()
-                                        : component;
-                                    return ButtonBuilder.from(payload).setDisabled(true);
-                                })
-                                .filter((btn: ButtonBuilder | null): btn is ButtonBuilder => btn !== null);
-                            if (disabledButtons.length === 0) return null;
-                            nextRow.addComponents(...disabledButtons);
-                            return nextRow;
-                        })
-                        .filter((row): row is ActionRowBuilder<ButtonBuilder> => row !== null);
-
+                    const disabledRows = disableAllButtons(interaction.message.components);
                     await interaction.editReply({ components: disabledRows }).catch(logger.error);
 
                     if (deps.handleRetry) {
